Fix troublesome hang if SSH sessions are maxed
This was failing because inside a file node (already using an SSH
session) it would try to load users/groups, which required another SSH
session. This fails if MaxSessions=1, or if we're doing huge stuff
maxing out the session count. To avoid this problem we just load the
user info for root before doing anything else.

Another issue is within the user/group loader, it was not releasing one
of the SSH handles until the end of the function call due to defer.
Calling Close() in the middle on the filehandle fixes that.